Introduce Conflicts to Book processing.
Summary:
* Use a synthetic PK for Book
* Add an Alias with scheme "PUB_ID" for each publisher id provided
* When creating aliases for a book, find dupes and create Conflicts.
* Future tool will find conflicts and allow user to merge, correct.

Details:
Using the publisher-provided id as our primary key seems to be a poor idea.
The publisher id is essentially user input and user input shouldn't be part
of our primary key. So we change the model to use the default Django PK
mechanism and treat the publisher's id simply as another alias.

The thinking here is that conflicts will happen. Let's make resolution
of conflicts a low-cost event. Doing so will require structuring some
data around a conflict and then writing a tool that can process this data.
At first this tool will be used by humans; as we gain insight we can add
some automated conflict processing.