Deal with SQLite's max_length issue. More tests.
SQLite doesn't have a maximum size for varchar.
http://www.sqlite.org/faq.html#q9

Hence, it does not heed max_length as defined on the CharFields.
I'm assuming that we'll eventually want to use a different
database that does require max_length.

Calling `full_clean()` is a bit of hack, but it seems like the
most unobtrusive hack at the moment. Of course, if we WANT to
always accept whatever data size we are given, great. This is
easy to reverse.